Added multiple text field support.
Removed special cases for Description and Text/Body field: these are
handled the same as other text fields now.  TextLine fields and
StringFields are ignored, except for the Title field.

We add an Archetypes and Dexterity type with several text fields for testing this.
We use the Archetypes type when running the tests on Plone 4.
We use the Dexterity type when running the tests on Plone 5.
It should work fine the other way around too, but this is easiest to test.

To test the types manually, you need to load testing.zcml.